move json format into package
