+++ pkg/app/context_ls.go
@@ -23,6 +23,8 @@ import (
 	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli/v2"
 
+	"github.com/tensorchord/envd/pkg/app/formatter"
+	"github.com/tensorchord/envd/pkg/app/formatter/json"
 	"github.com/tensorchord/envd/pkg/home"
 	"github.com/tensorchord/envd/pkg/types"
 )
@@ -31,22 +33,7 @@ var CommandContextList = &cli.Command{
 	Name:  "ls",
 	Usage: "List envd contexts",
 	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name:     "format",
-			Usage:    "Format of output, could be \"json\" or \"table\"",
-			Aliases:  []string{"f"},
-			Value:    "table",
-			Required: false,
-			Action: func(clicontext *cli.Context, v string) error {
-				switch v {
-				case
-					"table",
-					"json":
-					return nil
-				}
-				return errors.Errorf("Argument format only allows \"json\" and \"table\", found %v", v)
-			},
-		},
+		&formatter.FormatFlag,
 	},
 	Action: contextList,
 }
@@ -59,14 +46,14 @@ func contextList(clicontext *cli.Context) error {
 	format := clicontext.String("format")
 	switch format {
 	case "table":
-		renderTableContext(os.Stdout, contexts)
+		renderContext(os.Stdout, contexts)
 	case "json":
-		return renderJsonContext(contexts)
+		return json.PrintContext(contexts)
 	}
 	return nil
 }
 
-func renderTableContext(w io.Writer, contexts types.EnvdContext) {
+func renderContext(w io.Writer, contexts types.EnvdContext) {
 	table := tablewriter.NewWriter(w)
 	table.SetHeader([]string{"context", "builder", "builder addr", "runner", "runner addr"})
 
@@ -98,31 +85,4 @@ func renderTableContext(w io.Writer, contexts types.EnvdContext) {
 		table.Append(envRow)
 	}
 	table.Render()
-}
-
-type contextJsonDisplay struct {
-	Context     string `json:"context"`
-	Builder     string `json:"builder"`
-	BuilderAddr string `json:"builder_addr"`
-	Runner      string `json:"runner"`
-	RunnerAddr  string `json:"runner_addr,omitempty"`
-	Current     bool   `json:"current"`
-}
-
-func renderJsonContext(contexts types.EnvdContext) error {
-	output := []contextJsonDisplay{}
-	for _, p := range contexts.Contexts {
-		item := contextJsonDisplay{
-			Context:     p.Name,
-			Builder:     string(p.Builder),
-			BuilderAddr: fmt.Sprintf("%s://%s", p.Builder, p.BuilderAddress),
-			Runner:      string(p.Runner),
-			Current:     p.Name == contexts.Current,
-		}
-		if p.RunnerAddress != nil {
-			item.RunnerAddr = *p.RunnerAddress
-		}
-		output = append(output, item)
-	}
-	return PrintJson(output)
 }
