+++ pkg/app/env_describe.go
@@ -23,6 +23,8 @@ import (
 	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli/v2"
 
+	"github.com/tensorchord/envd/pkg/app/formatter"
+	"github.com/tensorchord/envd/pkg/app/formatter/json"
 	"github.com/tensorchord/envd/pkg/envd"
 	"github.com/tensorchord/envd/pkg/home"
 	"github.com/tensorchord/envd/pkg/types"
@@ -47,22 +49,7 @@ var CommandDescribeEnvironment = &cli.Command{
 			Aliases: []string{"e"},
 			Value:   getCurrentDirOrPanic(),
 		},
-		&cli.StringFlag{
-			Name:     "format",
-			Usage:    "Format of output, could be \"json\" or \"table\", could be \"json\" or \"table\"",
-			Aliases:  []string{"p"},
-			Value:    "table",
-			Required: false,
-			Action: func(clicontext *cli.Context, v string) error {
-				switch v {
-				case
-					"table",
-					"json":
-					return nil
-				}
-				return errors.Errorf("Argument format only allows \"json\" and \"table\", found %v", v)
-			},
-		},
+		&formatter.FormatFlag,
 	},
 	Action: getEnvironmentDescriptions,
 }
@@ -93,10 +80,10 @@ func getEnvironmentDescriptions(clicontext *cli.Context) error {
 	format := clicontext.String("format")
 	switch format {
 	case "table":
-		renderTableDependencies(os.Stdout, dep)
-		renderTablePortBindings(os.Stdout, ports)
+		renderDependencies(os.Stdout, dep)
+		renderPortBindings(os.Stdout, ports)
 	case "json":
-		return renderjsonEnvDesp(dep, ports)
+		return json.PrintEnvironmentDescriptions(dep, ports)
 	}
 
 	return nil
@@ -121,7 +108,7 @@ func createTable(w io.Writer, headers []string) *tablewriter.Table {
 	return table
 }
 
-func renderTablePortBindings(w io.Writer, ports []types.PortBinding) {
+func renderPortBindings(w io.Writer, ports []types.PortBinding) {
 	if ports == nil {
 		return
 	}
@@ -138,7 +125,7 @@ func renderTablePortBindings(w io.Writer, ports []types.PortBinding) {
 	table.Render()
 }
 
-func renderTableDependencies(w io.Writer, dep *types.Dependency) {
+func renderDependencies(w io.Writer, dep *types.Dependency) {
 	if dep == nil {
 		return
 	}
@@ -156,51 +143,4 @@ func renderTableDependencies(w io.Writer, dep *types.Dependency) {
 		table.Append(envRow)
 	}
 	table.Render()
-}
-
-type envJsonDescribe struct {
-	Ports        []envJsonPort       `json:"ports,omitempty"`
-	Dependencies []envJsonDependency `json:"dependencies,omitempty"`
-}
-type envJsonPort struct {
-	Name          string `json:"name"`
-	ContainerPort string `json:"container_port"`
-	Protocol      string `json:"protocol"`
-	HostIP        string `json:"host_ip"`
-	HostPort      string `json:"host_port"`
-}
-
-type envJsonDependency struct {
-	Name string `json:"name"`
-	Type string `json:"type"`
-}
-
-func renderjsonEnvDesp(dep *types.Dependency, ports []types.PortBinding) error {
-	output := envJsonDescribe{}
-
-	for _, port := range ports {
-		port := envJsonPort{
-			Name:          port.Name,
-			ContainerPort: port.Port,
-			Protocol:      port.Protocol,
-			HostIP:        port.HostIP,
-			HostPort:      port.HostPort,
-		}
-		output.Ports = append(output.Ports, port)
-	}
-	for _, p := range dep.PyPIPackages {
-		dependency := envJsonDependency{
-			Name: p,
-			Type: "Python",
-		}
-		output.Dependencies = append(output.Dependencies, dependency)
-	}
-	for _, p := range dep.APTPackages {
-		dependency := envJsonDependency{
-			Name: p,
-			Type: "APT",
-		}
-		output.Dependencies = append(output.Dependencies, dependency)
-	}
-	return PrintJson(output)
 }
