+++ pkg/app/env_ls.go
@@ -19,12 +19,13 @@ import (
 	"io"
 	"os"
 	"strconv"
-	"strings"
 
 	"github.com/cockroachdb/errors"
 	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli/v2"
 
+	"github.com/tensorchord/envd/pkg/app/formatter"
+	"github.com/tensorchord/envd/pkg/app/formatter/json"
 	"github.com/tensorchord/envd/pkg/app/telemetry"
 	"github.com/tensorchord/envd/pkg/envd"
 	"github.com/tensorchord/envd/pkg/home"
@@ -36,22 +37,7 @@ var CommandListEnv = &cli.Command{
 	Aliases: []string{"ls", "l"},
 	Usage:   "List envd environments",
 	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name:     "format",
-			Usage:    "Format of output, could be \"json\" or \"table\"",
-			Aliases:  []string{"f"},
-			Value:    "table",
-			Required: false,
-			Action: func(clicontext *cli.Context, v string) error {
-				switch v {
-				case
-					"table",
-					"json":
-					return nil
-				}
-				return errors.Errorf("Argument format only allows \"json\" and \"table\", found %v", v)
-			},
-		},
+		&formatter.FormatFlag,
 	},
 	Action: getEnvironment,
 }
@@ -77,14 +63,14 @@ func getEnvironment(clicontext *cli.Context) error {
 	format := clicontext.String("format")
 	switch format {
 	case "table":
-		renderTableEnvironments(os.Stdout, envs)
+		renderEnvironments(os.Stdout, envs)
 	case "json":
-		return renderjsonEnvironments(envs)
+		return json.PrintEnvironments(envs)
 	}
 	return nil
 }
 
-func renderTableEnvironments(w io.Writer, envs []types.EnvdEnvironment) {
+func renderEnvironments(w io.Writer, envs []types.EnvdEnvironment) {
 	table := tablewriter.NewWriter(w)
 	table.SetHeader([]string{
 		"Name", "Endpoint", "SSH Target", "Image",
@@ -106,7 +92,7 @@ func renderTableEnvironments(w io.Writer, envs []types.EnvdEnvironment) {
 	for _, env := range envs {
 		envRow := make([]string, 9)
 		envRow[0] = env.Name
-		envRow[1] = endpointOrNone(env)
+		envRow[1] = formatter.FormatEndpoint(env)
 		envRow[2] = fmt.Sprintf("%s.envd", env.Name)
 		envRow[3] = env.Spec.Image
 		envRow[4] = strconv.FormatBool(env.GPU)
@@ -116,44 +102,4 @@ func renderTableEnvironments(w io.Writer, envs []types.EnvdEnvironment) {
 		table.Append(envRow)
 	}
 	table.Render()
-}
-
-type envJsonDisplay struct {
-	Name      string `json:"name"`
-	Endpoint  string `json:"endpoint,omitempty"`
-	SSHTarget string `json:"ssh_target"`
-	Image     string `json:"image"`
-	GPU       bool   `json:"gpu"`
-	CUDA      string `json:"cuda,omitempty"`
-	CUDNN     string `json:"cudnn,omitempty"`
-	Status    string `json:"status"`
-}
-
-func renderjsonEnvironments(envs []types.EnvdEnvironment) error {
-	output := []envJsonDisplay{}
-	for _, env := range envs {
-		item := envJsonDisplay{
-			Name:      env.Name,
-			Endpoint:  endpointOrNone(env),
-			SSHTarget: fmt.Sprintf("%s.envd", env.Name),
-			Image:     env.Spec.Image,
-			GPU:       env.GPU,
-			CUDA:      env.CUDA,
-			CUDNN:     env.CUDNN,
-			Status:    env.Status.Phase,
-		}
-		output = append(output, item)
-	}
-	return PrintJson(output)
-}
-
-func endpointOrNone(env types.EnvdEnvironment) string {
-	var res strings.Builder
-	if env.Status.JupyterAddr != nil {
-		res.WriteString(fmt.Sprintf("jupyter: %s", *env.Status.JupyterAddr))
-	}
-	if env.Status.RStudioServerAddr != nil {
-		res.WriteString(fmt.Sprintf("rstudio: %s", *env.Status.RStudioServerAddr))
-	}
-	return res.String()
 }
