+++ pkg/app/image.go
@@ -26,6 +26,8 @@ import (
 	"github.com/olekukonko/tablewriter"
 	"github.com/urfave/cli/v2"
 
+	"github.com/tensorchord/envd/pkg/app/formatter"
+	"github.com/tensorchord/envd/pkg/app/formatter/json"
 	"github.com/tensorchord/envd/pkg/envd"
 	"github.com/tensorchord/envd/pkg/home"
 	"github.com/tensorchord/envd/pkg/types"
@@ -50,22 +52,7 @@ var CommandListImage = &cli.Command{
 	Aliases: []string{"ls", "l"},
 	Usage:   "List envd images",
 	Flags: []cli.Flag{
-		&cli.StringFlag{
-			Name:     "format",
-			Usage:    "Format of output, could be \"json\" or \"table\"",
-			Aliases:  []string{"f"},
-			Value:    "table",
-			Required: false,
-			Action: func(clicontext *cli.Context, v string) error {
-				switch v {
-				case
-					"table",
-					"json":
-					return nil
-				}
-				return errors.Errorf("Argument format only allows \"json\" and \"table\", found %v", v)
-			},
-		},
+		&formatter.FormatFlag,
 	},
 	Action: getImage,
 }
@@ -89,14 +76,14 @@ func getImage(clicontext *cli.Context) error {
 	format := clicontext.String("format")
 	switch format {
 	case "table":
-		renderTableImages(os.Stdout, imgs)
+		renderImages(os.Stdout, imgs)
 	case "json":
-		return renderjsonImages(imgs)
+		return json.PrintImages(imgs)
 	}
 	return nil
 }
 
-func renderTableImages(w io.Writer, imgs []types.EnvdImage) {
+func renderImages(w io.Writer, imgs []types.EnvdImage) {
 	table := tablewriter.NewWriter(w)
 	table.SetHeader([]string{"Name", "Context", "GPU", "CUDA", "CUDNN", "Image ID", "Created", "Size"})
 
@@ -125,35 +112,6 @@ func renderTableImages(w io.Writer, imgs []types.EnvdImage) {
 		table.Append(envRow)
 	}
 	table.Render()
-}
-
-type imgJsonDisplay struct {
-	Name    string `json:"name"`
-	Context string `json:"endpoint,omitempty"`
-	GPU     bool   `json:"gpu"`
-	CUDA    string `json:"cuda,omitempty"`
-	CUDNN   string `json:"cudnn,omitempty"`
-	ImageID string `json:"image_id"`
-	Created string `json:"created"`
-	Size    string `json:"size"`
-}
-
-func renderjsonImages(imgs []types.EnvdImage) error {
-	output := []imgJsonDisplay{}
-	for _, img := range imgs {
-		item := imgJsonDisplay{
-			Name:    img.Name,
-			Context: img.BuildContext,
-			GPU:     img.GPU,
-			CUDA:    img.CUDA,
-			CUDNN:   img.CUDNN,
-			ImageID: stringid.TruncateID(img.Digest),
-			Created: createdSinceString(img.Created),
-			Size:    units.HumanSizeWithPrecision(float64(img.Size), 3),
-		}
-		output = append(output, item)
-	}
-	return PrintJson(output)
 }
 
 func stringOrNone(cuda string) string {
